Move VI model constants to `tools/_core.py`
